refactor: decompose 27-plan-run-plot-roundtrip.ts and lower file-size budget (warren-5c16)
