+++ .mulch/expertise/acceptance.jsonl
@@ -37,3 +37,4 @@
 {"type":"convention","classification":"tactical","recorded_at":"2026-05-19T23:50:24.352Z","evidence":{"seeds":"warren-73ee","commit":"c88907c76e28fc85cebf3bea9740b1ea80ffb288"},"relates_to":["mx-629c78"],"content":"Scenario 25 (scripts/acceptance/scenarios/25-plot-roundtrip.ts, warren-73ee / pl-95dd step 3) verified green end-to-end against a live in-proc warren+burrow stack with @os-eco/burrow-cli@0.3.3 (the burrow-59cd release). All four previously soft-skipped Plot assertions — PLOT_ID echo, PLOT_ACTOR echo, agent plot-append OK, reap mirror of plot.decision_made — run unconditionally and pass. Sibling scenario 27 (PlanRun + Plot) also green so the body.env parse fix did not regress the parallel plan-run path. Run command: 'bun run scripts/acceptance/run.ts --only 25'. Typical wall time ~4.6s. macOS works (in-proc only, modes: ['in-proc']).","id":"mx-0f1e6d"}
 {"type":"pattern","classification":"tactical","recorded_at":"2026-05-28T04:27:36.462Z","evidence":{"seeds":"warren-69fe","commit":"e46f2af49b7779fdaa04ec8d7b122097768bf76c"},"name":"Decomposing sequential acceptance scenarios into scenarios/lib siblings","description":"When an acceptance scenario file grows beyond the 500-line budget, it should be decomposed by extracting shared setup, assertion, and request-builder helpers into modular sibling files under scripts/acceptance/scenarios/lib/. This keeps the scenario orchestrator thin, legible, and easy to maintain while enabling reusability.","files":["scripts/acceptance/scenarios/29-plot-detail-roundtrip.ts","scripts/acceptance/scenarios/lib/types.ts","scripts/acceptance/scenarios/lib/git-helpers.ts","scripts/acceptance/scenarios/lib/poll-helpers.ts","scripts/acceptance/scenarios/lib/event-helpers.ts","scripts/acceptance/scenarios/lib/fixture-29.ts"],"id":"mx-229651"}
 {"type":"convention","classification":"tactical","recorded_at":"2026-05-28T04:34:28.519Z","evidence":{"seeds":"warren-c83b","commit":"9a0879793393298aab054685d522659899b829f2"},"content":"Decomposed scripts/acceptance/scenarios/31-plot-plan-run-synthesis.ts (950 lines) by extracting shared synthesis fixtures, build helpers, and git config redirect writers into scripts/acceptance/scenarios/lib/fixture-31.ts; leveraged existing poll-helpers and event-helpers to reduce the orchestrator scenario file under 500 lines.","id":"mx-f05190"}
+{"type":"convention","classification":"tactical","recorded_at":"2026-05-28T04:41:55.522Z","evidence":{"commit":"0f61bfff54f444b0fb454e27ecd1fb07e7367db3"},"content":"Decomposed scripts/acceptance/scenarios/27-plan-run-plot-roundtrip.ts (856 lines) into scenarios/lib/fixture-27.ts and imported shared poll/event helpers, lowering file-size budget below 500-line ceiling.","id":"mx-e88e1d"}
