+++ scripts/acceptance/scenarios/27-plan-run-plot-roundtrip.ts
@@ -82,93 +82,36 @@
  * keep the scenario's fixture footprint to one project clone.
  */
 
-import { existsSync } from "node:fs";
-import { copyFile, mkdir, mkdtemp, readFile, writeFile } from "node:fs/promises";
+import { mkdtemp } from "node:fs/promises";
 import { tmpdir } from "node:os";
 import { join } from "node:path";
 
 import { AcceptanceError, assertEqual, assertTrue, type Scenario } from "../lib/assert.ts";
 import { WarrenHttp } from "../lib/http.ts";
 import { type BootHandle, bootInProc } from "../lib/inproc.ts";
+import {
+	fetchAllPlanRunEvents,
+	fetchAllRunEvents,
+	findTextEvent,
+	parsePlotLines,
+	readPlotEventLines,
+} from "./lib/event-helpers.ts";
+import {
+	buildPlanRunPlotFixture,
+	PLAN_ID_BASE,
+	PLAN_ID_PLOT,
+	PLAN_PROJECT_URL,
+	SEED_C,
+} from "./lib/fixture-27.ts";
+import {
+	readPlotSnapshot,
+	sleep,
+	waitForPlanState,
+	waitForPlotStatus,
+} from "./lib/poll-helpers.ts";
+import type { CreatePlanRunResponse, ProjectRow } from "./lib/types.ts";
 
-interface ProjectRow {
-	readonly id: string;
-	readonly gitUrl: string;
-	readonly localPath: string;
-	readonly defaultBranch: string;
-	readonly hasSeeds?: boolean;
-	readonly hasPlot?: boolean;
-}
-
-interface RunRow {
-	readonly id: string;
-	readonly state: string;
-	readonly trigger: string;
-	readonly prUrl: string | null;
-	readonly plotId: string | null;
-}
-
-interface PlanRunRow {
-	readonly id: string;
-	readonly planId: string;
-	readonly projectId: string;
-	readonly agentName: string;
-	readonly state: "queued" | "running" | "succeeded" | "failed" | "cancelled";
-	readonly plotId: string | null;
-}
-
-interface PlanRunChildRow {
-	readonly planRunId: string;
-	readonly seq: number;
-	readonly seedId: string;
-	readonly runId: string | null;
-	readonly state:
-		| "pending"
-		| "dispatched"
-		| "running"
-		| "pr_open"
-		| "merged"
-		| "failed"
-		| "skipped";
-}
-
-interface CreatePlanRunResponse {
-	readonly planRun: PlanRunRow;
-	readonly children: readonly PlanRunChildRow[];
-}
-
-interface PlanRunDetailResponse {
-	readonly planRun: PlanRunRow;
-	readonly children: readonly PlanRunChildRow[];
-	readonly runs: readonly RunRow[];
-}
-
-interface EventRow {
-	readonly id: number;
-	readonly runId: string;
-	readonly seq: number;
-	readonly kind: string;
-	readonly stream: string | null;
-	readonly payload: Record<string, unknown> | null;
-}
-
-interface PlotSnapshot {
-	readonly id: string;
-	readonly status: string;
-}
-
-const PLAN_PROJECT_URL = "https://github.com/warren-acceptance/sample-plan-run-plot.git";
-const PLAN_ID_PLOT = "pl-acc-27-plot";
-const PLAN_ID_BASE = "pl-acc-27-base";
-const SEED_A = "ah-acc-27-a";
-const SEED_B = "ah-acc-27-b";
-const SEED_C = "ah-acc-27-c";
-const SEED_D = "ah-acc-27-d";
-const SEED_TS = "2026-05-17T00:00:00.000Z";
-
-const TERMINAL_PLAN_STATES = new Set(["succeeded", "failed", "cancelled"]);
 const PLAN_DEADLINE_MS = 90_000;
-const POLL_INTERVAL_MS = 500;
 const PLOT_FILE_POLL_TIMEOUT_MS = 10_000;
 /**
  * After waitForPlanState returns, give the auto-done transition one more
@@ -544,313 +487,3 @@ export const scenario: Scenario = {
 		}
 	},
 };
-
-interface BuildPlanRunPlotFixtureInput {
-	readonly fixturePath: string;
-	readonly sourceSamplePath: string;
-	readonly harnessGitConfigPath: string;
-	readonly gitConfigPath: string;
-	readonly projectGitUrl: string;
-}
-
-/**
- * Build a fixture mirroring scenario 26's `.seeds/`-enabled layout plus a
- * committed `.plot/` directory holding one Plot pre-transitioned to
- * `active`. Returns the plot id so the scenario can dispatch with it.
- *
- * Two plans land in `.seeds/plans.jsonl` so the scenario can hit both
- * dispatch shapes against the same project without re-cloning.
- */
-async function buildPlanRunPlotFixture(input: BuildPlanRunPlotFixtureInput): Promise<string> {
-	await mkdir(input.fixturePath, { recursive: true });
-	await mkdir(join(input.fixturePath, "tools"), { recursive: true });
-	await mkdir(join(input.fixturePath, ".seeds"), { recursive: true });
-
-	const burrowToml = await readFile(join(input.sourceSamplePath, "burrow.toml"), "utf8");
-	await writeFile(join(input.fixturePath, "burrow.toml"), burrowToml);
-	await copyFile(
-		join(input.sourceSamplePath, "tools", "stub-agent.sh"),
-		join(input.fixturePath, "tools", "stub-agent.sh"),
-	);
-	await copyFile(
-		join(input.sourceSamplePath, "tools", "claude-code-stub-agent.sh"),
-		join(input.fixturePath, "tools", "claude-code-stub-agent.sh"),
-	);
-	await writeFile(
-		join(input.fixturePath, "README.md"),
-		"# warren acceptance plan-run + plot fixture\n\nUsed by scripts/acceptance/scenarios/27-plan-run-plot-roundtrip.ts.\n",
-	);
-
-	await writeFile(
-		join(input.fixturePath, ".seeds", "config.yaml"),
-		`project: "sample-plan-run-plot"\nversion: "1"\nmax_plan_depth: 3\n`,
-	);
-	await writeFile(
-		join(input.fixturePath, ".seeds", "issues.jsonl"),
-		[seedRowOpen(SEED_A), seedRowOpen(SEED_B), seedRowOpen(SEED_C), seedRowOpen(SEED_D)].join(""),
-	);
-	await writeFile(
-		join(input.fixturePath, ".seeds", "plans.jsonl"),
-		[planRow(PLAN_ID_PLOT, [SEED_A, SEED_B, SEED_C]), planRow(PLAN_ID_BASE, [SEED_D])].join(""),
-	);
-
-	const env = withGitIdentity();
-	await runIn(input.fixturePath, ["git", "init", "--initial-branch=main"], env);
-	await runIn(input.fixturePath, ["chmod", "+x", "tools/stub-agent.sh"], env);
-	await runIn(input.fixturePath, ["chmod", "+x", "tools/claude-code-stub-agent.sh"], env);
-
-	// `plot init` with a user actor (Plot SPEC §6 forbids agent actors on
-	// plot_created). Then transition drafting → ready → active so the
-	// coordinator's auto-done has an `active` plot to terminate. Without
-	// this the auto-done would correctly skip-with-currentStatus=drafting,
-	// which is a different code path (acceptance #9).
-	const plotEnv: Record<string, string> = { ...env, PLOT_ACTOR: "user:acceptance" };
-	await runIn(input.fixturePath, ["plot", "init", "scenario-27"], plotEnv);
-	const list = await runIn(input.fixturePath, ["plot", "list", "--json"], env);
-	const plots = JSON.parse(list.stdout) as ReadonlyArray<{ id: string }>;
-	if (plots.length !== 1) {
-		throw new AcceptanceError(
-			`scenario-27 fixture: expected exactly one Plot after init, got ${plots.length}: ${list.stdout}`,
-		);
-	}
-	const plotId = plots[0]?.id;
-	if (plotId === undefined) {
-		throw new AcceptanceError(`scenario-27 fixture: plot list --json missing id: ${list.stdout}`);
-	}
-	await runIn(input.fixturePath, ["plot", "status", plotId, "ready"], plotEnv);
-	await runIn(input.fixturePath, ["plot", "status", plotId, "active"], plotEnv);
-
-	await runIn(input.fixturePath, ["git", "add", "."], env);
-	await runIn(
-		input.fixturePath,
-		["git", "commit", "-m", "init: plan-run + plot acceptance fixture"],
-		env,
-	);
-
-	const harnessConfig = existsSync(input.harnessGitConfigPath)
-		? await readFile(input.harnessGitConfigPath, "utf8")
-		: "";
-	const lines: string[] = [
-		harnessConfig.trimEnd(),
-		`[url "${input.fixturePath}"]`,
-		`\tinsteadOf = ${input.projectGitUrl}`,
-		"",
-	];
-	await writeFile(input.gitConfigPath, `${lines.join("\n")}\n`);
-
-	return plotId;
-}
-
-function seedRowOpen(id: string): string {
-	const row = {
-		id,
-		title: `scenario-27 ${id}`,
-		status: "open",
-		type: "task",
-		priority: 3,
-		createdAt: SEED_TS,
-		updatedAt: SEED_TS,
-	};
-	return `${JSON.stringify(row)}\n`;
-}
-
-function planRow(id: string, children: readonly string[]): string {
-	const plan = {
-		id,
-		seed: "warren-acc-27",
-		template: "feature",
-		status: "approved",
-		revision: 1,
-		sections: {
-			context: `scenario-27 acceptance plan ${id}`,
-			approach: "dispatch child seeds via the plan-run coordinator",
-			steps: children.map((s) => ({ title: `close ${s}` })),
-		},
-		children,
-		createdAt: SEED_TS,
-		updatedAt: SEED_TS,
-		name: `scenario-27 ${id}`,
-	};
-	return `${JSON.stringify(plan)}\n`;
-}
-
-async function waitForPlanState(
-	http: WarrenHttp,
-	planRunId: string,
-	target: string,
-	timeoutMs: number,
-): Promise<PlanRunDetailResponse> {
-	const start = Date.now();
-	let last = "unknown";
-	while (Date.now() - start < timeoutMs) {
-		const row = await http.expectJson<PlanRunDetailResponse>(
-			"GET",
-			`/plan-runs/${encodeURIComponent(planRunId)}`,
-			200,
-		);
-		last = row.planRun.state;
-		if (row.planRun.state === target) return row;
-		if (TERMINAL_PLAN_STATES.has(row.planRun.state)) {
-			throw new AcceptanceError(
-				`plan-run ${planRunId}: expected '${target}', reached terminal '${row.planRun.state}'`,
-			);
-		}
-		await sleep(POLL_INTERVAL_MS);
-	}
-	throw new AcceptanceError(
-		`plan-run ${planRunId} did not reach '${target}' within ${timeoutMs}ms (last state=${last})`,
-	);
-}
-
-async function fetchAllPlanRunEvents(http: WarrenHttp, planRunId: string): Promise<EventRow[]> {
-	const events: EventRow[] = [];
-	for await (const row of http.streamNdjson(`/plan-runs/${encodeURIComponent(planRunId)}/events`)) {
-		events.push(row as EventRow);
-	}
-	return events;
-}
-
-async function fetchAllRunEvents(http: WarrenHttp, runId: string): Promise<EventRow[]> {
-	const events: EventRow[] = [];
-	for await (const row of http.streamNdjson(`/runs/${encodeURIComponent(runId)}/events`)) {
-		events.push(row as EventRow);
-	}
-	return events;
-}
-
-function findTextEvent(events: readonly EventRow[], needle: string): EventRow | undefined {
-	return events.find(
-		(e) =>
-			e.kind === "text" &&
-			typeof e.payload?.text === "string" &&
-			(e.payload.text as string).includes(needle),
-	);
-}
-
-/**
- * Read the Plot events.jsonl from disk and return the set of trimmed
- * non-empty lines. As of warren-fdd2 (pl-d4d6 step 1) refreshProjectClone
- * snapshots `.plot/` out-of-tree before `git reset --hard` and restores
- * it after, so host-side appender writes accumulate across child
- * dispatches and a single post-completion read sees every line. Missing
- * file is treated as an empty set so the helper composes with the
- * pre-dispatch baseline snapshot.
- */
-async function readPlotEventLines(path: string): Promise<ReadonlySet<string>> {
-	const seen = new Set<string>();
-	try {
-		const body = await readFile(path, "utf8");
-		for (const line of body.split("\n")) {
-			const trimmed = line.trim();
-			if (trimmed === "") continue;
-			seen.add(trimmed);
-		}
-	} catch {
-		// File not yet present — return empty set.
-	}
-	return seen;
-}
-
-interface ParsedPlotEvent {
-	readonly type: string;
-	readonly actor: string;
-	readonly at: string;
-	readonly data: unknown;
-}
-
-function parsePlotLines(lines: ReadonlySet<string>): ParsedPlotEvent[] {
-	const out: ParsedPlotEvent[] = [];
-	for (const line of lines) {
-		let parsed: unknown;
-		try {
-			parsed = JSON.parse(line);
-		} catch {
-			continue;
-		}
-		if (typeof parsed !== "object" || parsed === null) continue;
-		const row = parsed as { type?: unknown; actor?: unknown; at?: unknown; data?: unknown };
-		if (
-			typeof row.type !== "string" ||
-			typeof row.actor !== "string" ||
-			typeof row.at !== "string"
-		) {
-			continue;
-		}
-		out.push({ type: row.type, actor: row.actor, at: row.at, data: row.data ?? null });
-	}
-	return out;
-}
-
-async function readPlotSnapshot(path: string): Promise<PlotSnapshot> {
-	const body = await readFile(path, "utf8");
-	const parsed = JSON.parse(body) as PlotSnapshot;
-	return parsed;
-}
-
-async function waitForPlotStatus(
-	path: string,
-	target: string,
-	timeoutMs: number,
-): Promise<PlotSnapshot> {
-	const start = Date.now();
-	let lastStatus = "unknown";
-	while (Date.now() - start < timeoutMs) {
-		try {
-			const snapshot = await readPlotSnapshot(path);
-			lastStatus = snapshot.status;
-			if (snapshot.status === target) return snapshot;
-		} catch {
-			// File not yet present or mid-write — keep polling.
-		}
-		await sleep(100);
-	}
-	throw new AcceptanceError(
-		`Plot at ${path} did not reach status='${target}' within ${timeoutMs}ms (last status=${lastStatus})`,
-	);
-}
-
-interface RunResult {
-	stdout: string;
-	stderr: string;
-}
-
-async function runIn(
-	cwd: string,
-	cmd: readonly string[],
-	env: Record<string, string>,
-): Promise<RunResult> {
-	const proc = Bun.spawn({
-		cmd: [...cmd],
-		cwd,
-		env,
-		stdin: "ignore",
-		stdout: "pipe",
-		stderr: "pipe",
-	});
-	const [stdout, stderr, exitCode] = await Promise.all([
-		new Response(proc.stdout).text(),
-		new Response(proc.stderr).text(),
-		proc.exited,
-	]);
-	if ((exitCode ?? 0) !== 0) {
-		throw new AcceptanceError(
-			`scenario-27 command failed (${cmd.join(" ")} in ${cwd}): exit ${exitCode}\nstderr: ${stderr}\nstdout: ${stdout}`,
-		);
-	}
-	return { stdout, stderr };
-}
-
-function withGitIdentity(): Record<string, string> {
-	return {
-		PATH: process.env.PATH ?? "",
-		HOME: process.env.HOME ?? "/tmp",
-		GIT_AUTHOR_NAME: "Warren Acceptance",
-		GIT_AUTHOR_EMAIL: "acceptance@warren.invalid",
-		GIT_COMMITTER_NAME: "Warren Acceptance",
-		GIT_COMMITTER_EMAIL: "acceptance@warren.invalid",
-	};
-}
-
-function sleep(ms: number): Promise<void> {
-	return new Promise((resolve) => setTimeout(resolve, ms));
-}
